DRY-ed up the tests, remove extraneous activity
