+++ app/src/main/java/com/h/pixeldroid/fragments/FeedRecyclerViewAdapter.kt
@@ -1,17 +1,15 @@
-package com.h.pixeldroid
+package com.h.pixeldroid.fragments
 
 import android.content.Context
 import android.graphics.Typeface
-import android.graphics.drawable.Drawable
 import android.util.DisplayMetrics
 import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.view.WindowManager
 import android.widget.ImageView
 import android.widget.TextView
-import androidx.annotation.Dimension
+import com.h.pixeldroid.R
 import com.h.pixeldroid.models.Post
 import com.h.pixeldroid.utils.ImageConverter.Companion.setDefaultImage
 import com.h.pixeldroid.utils.ImageConverter.Companion.setImageViewFromURL
