--- app/src/main/res/layout/activity_feed.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".FeedActivity">
-
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/feedList"
-        android:name="com.h.pixeldroid.FeedFragment"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_marginLeft="16dp"
-        android:layout_marginRight="16dp"
-        app:layoutManager="LinearLayoutManager"
-        tools:context=".FeedActivity"
-        tools:listitem="@layout/post_fragment" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
