+++ app/src/main/res/layout/fragment_feed_list.xml
@@ -9,5 +9,4 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".FeedActivity"
     tools:listitem="@layout/post_fragment" />
