+++ app/src/main/res/layout/fragment_home.xml
@@ -14,7 +14,6 @@
         android:layout_marginLeft="16dp"
         android:layout_marginRight="16dp"
         app:layoutManager="LinearLayoutManager"
-        tools:context=".FeedActivity"
         tools:listitem="@layout/post_fragment" />
 
     <LinearLayout
